+++ synergy-fitness/src/app/app-routing.module.ts
@@ -10,6 +10,7 @@ import { SignupComponent } from './components/signup/signup.component';
 import { UploadComponent } from './components/upload/upload.component';
 import { UserprofileComponent } from './components/userprofile/userprofile.component';
 import { ProfilesComponent } from './components/profiles/profiles.component';
+import { CreatePostComponent } from './tools/create-post/create-post.component';
 
 const routes: Routes = [
   { path:'', component: HomeComponent },
@@ -21,7 +22,8 @@ const routes: Routes = [
   { path: 'post', component: PostComponent },
   { path: 'upload', component: UploadComponent },
   { path: 'userprofile', component: UserprofileComponent },
-  { path: 'userprofile', component: UserprofileComponent }
+  { path: 'userprofile', component: UserprofileComponent },
+  { path: 'createpost', component: CreatePostComponent }
 ];
 
 @NgModule({
