+++ synergy-fitness/src/app/components/header/header.component.html
@@ -11,7 +11,7 @@
       <div class="collapse navbar-collapse" id="navbarSupportedContent">
         <ul class="navbar-nav me-auto mb-2 mb-lg-0">
           <li class="nav-item">
-            <a class="nav-link" aria-current="page" [routerLink]="'/'" [routerLinkActive]="'active'"
+            <a class="nav-link" aria-current="page" [routerLink]="'/post'" [routerLinkActive]="'active'"
               [routerLinkActiveOptions]="{ exact: true }">Home</a>
           </li>
           <li class="nav-item">
