+++ synergy-fitness/src/app/components/home/home.component.html
@@ -1,7 +1,7 @@
 <mat-card>
     <mat-card-header>
-        <mat-card-title>Name</mat-card-title>
-        <mat-card-subtitle>Description</mat-card-subtitle>
+        <mat-card-title>{{this.getpost.user.gymUsername}}</mat-card-title>
+        <mat-card-subtitle>{{this.getpost.postData}}</mat-card-subtitle>
     </mat-card-header>
     <!--<img mat-card-image src="C:\Users\user\Documents\Workouts\workout2.jpg">-->
     <mat-card-content>
@@ -13,6 +13,7 @@
         <mat-icon class="action-item" color="warn">save</mat-icon>
     </mat-card-actions>
 </mat-card>
+
 
 
 <!--<h1>This is Home page</h1>
