+++ synergy-fitness/src/app/models/post.ts
@@ -3,7 +3,7 @@ import { Person } from "./person";
 export class Post {
     constructor(
         public postId:number,
-        public postUser:Person,
+        public user:Person,
         public postData:string
     ){}
 }
